requirements.txt
streamlit==1.35.0
gspread==6.0.2
pandas==2.2.2
google-auth==2.29.0
